Code cleaning and bug fixes, prepping for Edit Modal functionality
Moved the modal validation to it's own function so the same function can
be used to validate the edit modal.

Update the edit modal to work like the add modal.

Fixed a bug with the recipe id's not being set correctly

Next update should add functionality to the edit modal by updating the
application state's edit object to represent the same recipe that the
user is trying to edit